@@ -1,88 +1,90 @@
 {
-  "name": "gsoc2",
-  "homepage": "https://soc2.khulnasoft.com",
-  "version": "1.2.0",
+  "name": "gsoc2r",
+  "homepage": "https://gsoc2r.io",
+  "version": "1.3.0",
   "private": true,
   "dependencies": {
-    "@codemirror/commands": "^6.2.2",
+    "@codemirror/commands": "^6.2.4",
     "@emotion/is-prop-valid": "^1.1.1",
-    "@emotion/react": "^11.7.0",
-    "@emotion/styled": "^11.6.0",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
     "@lezer/highlight": "^1.1.3",
-    "@material-ui/core": "^4.5.2",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/lab": "^4.0.0-alpha.58",
-    "@material-ui/styles": "^4.5.2",
-    "@material-ui/utils": "^4.11.2",
-    "@metamask/detect-provider": "^1.2.0",
-    "@mui/icons-material": "^5.2.1",
-    "@mui/material": "^5.2.3",
+    "@mui/icons-material": "^5.14.0",
+    "@mui/material": "^5.14.0",
+    "@mui/styles": "^5.14.0",
     "@mui/x-data-grid": "^5.17.11",
-    "@uiw/codemirror-themes": "^4.19.9",
-    "@uiw/react-codemirror": "^3.2.1",
+    "@mui/x-date-pickers": "^6.11.1",
+    "@uiw/codemirror-theme-vscode": "^4.21.20",
+    "@uiw/codemirror-themes": "^4.21.9",
+    "@uiw/react-codemirror": "^4.21.9",
     "@use-it/interval": "^1.0.0",
     "algoliasearch": "^4.13.1",
+    "calculate-size": "^1.1.1",
     "class-transformer": "^0.4.0",
-    "create-react-app": "^4.0.3",
+    "create-react-app": "^5.0.1",
     "cytoscape": "^3.15.1",
-    "cytoscape-clipboard": "^2.2.1",
-    "cytoscape-cxtmenu": "^3.1.1",
     "cytoscape-edgehandles": "^3.6.0",
-    "cytoscape-grid-guide": "~2.3.3",
     "cytoscape-node-html-label": "^1.1.5",
-    "cytoscape-panzoom": "^2.5.2",
-    "cytoscape-undo-redo": "^1.3.2",
     "d3": "^7.1.1",
+    "dayjs": "^1.11.9",
     "dotenv": "^6.1.0",
     "downshift": "^3.3.5",
     "github-markdown-css": "^3.0.1",
     "import": "0.0.6",
     "interweave": "^11.2.0",
-    "material-icons": "^0.7.7",
-    "material-icons-react": "^1.0.4",
-    "material-ui-chip-input": "^2.0.0-beta.2",
-    "material-ui-nested-menu-item": "^1.0.2",
+    "jss": "^10.10.0",
+    "jss-camel-case": "^6.1.0",
+    "jss-default-unit": "^8.0.2",
+    "jss-global": "^3.0.0",
+    "jss-nested": "^6.0.1",
+    "jss-props-sort": "^6.0.0",
+    "jss-vendor-prefixer": "^8.0.1",
     "md5-file": "^4.0.0",
     "mdbreact": "^4.21.1",
+    "mime": "^3.0.0",
     "moment": "^2.29.1",
-    "react": "^16.14.0",
-    "react-alert": "^5.5.0",
+    "mui-chips-input": "^2.1.3",
+    "mui-nested-menu": "^3.2.1",
+    "process": "^0.11.10",
+    "react": "^18.2.0",
+    "react-alert": "^7.0.3",
     "react-alert-template-basic": "^1.0.0",
     "react-alice-carousel": "^2.6.4",
     "react-avatar-editor": "^11.1.0",
     "react-beforeunload": "^2.2.1",
     "react-chartjs-2": "^2.11.1",
     "react-cookie": "^4.0.1",
-    "react-cytoscapejs": "^1.2.0",
-    "react-device-detect": "^1.9.10",
-    "react-dom": "^16.14.0",
+    "react-cytoscapejs": "^2.0.0",
+    "react-device-detect": "^2.2.3",
+    "react-dom": "^18.2.0",
     "react-draggable": "^3.3.2",
     "react-driftjs": "^1.2.2",
-    "react-dropzone": "^10.1.10",
+    "react-dropzone": "^14.2.3",
     "react-ga4": "^2.0.0",
-    "react-iframe": "^1.8.0",
+    "react-hotkeys": "^2.0.0",
     "react-instantsearch-dom": "^6.28.0",
     "react-json-pretty": "^2.2.0",
-    "react-json-view": "^1.19.1",
-    "react-markdown": "^4.2.2",
+    "react-json-view": "^1.21.3",
+    "react-markdown": "^8.0.7",
     "react-markdown-github": "^3.3.1",
-    "react-powerhooks": "0.0.7",
-    "react-router": "6.2.1",
-    "react-router-dom": "6.2.1",
-    "react-scripts": "^4.0.1",
-    "react-shepherd": "^3.3.6",
-    "reactstrap": "^7.1.0",
-    "reaviz": "^12.1.0",
+    "react-powerhooks": "^0.0.7",
+    "react-router": "^6.14.1",
+    "react-router-dom": "^6.14.1",
+    "react-scripts": "^5.0.1",
+    "react-toastify": "^9.1.3",
+    "reaviz": "^14.9.4",
+    "remark-gfm": "^3.0.1",
     "search-insights": "^2.2.1",
     "shellwords": "^0.1.1",
     "simplebar": "^4.2.3",
     "styled-components": "^4.4.0",
+    "webpack": "^5.88.2",
     "yaml": "^1.7.2",
     "yamljs": "^0.3.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "^0.13.1"
   },
   "scripts": {
-    "start": "HTTPS=false&&PORT=3000 react-scripts --openssl-legacy-provider start",
+    "start": "HTTPS=false&&PORT=3000 GENERATE_SOURCEMAP=false react-scripts --openssl-legacy-provider start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
