@@ -1,9 +1,14 @@
 import React from "react";
-import InfoIcon from "@material-ui/icons/Info";
-import CheckIcon from "@material-ui/icons/Check";
-import ErrorOutlineIcon from "@material-ui/icons/ErrorOutline";
-import CloseIcon from "@material-ui/icons/Close";
-import Typography from "@material-ui/core/Typography";
+import {
+	Info as InfoIcon, 
+	Check as CheckIcon, 
+	ErrorOutline as ErrorOutlineIcon, 
+	Close as CloseIcon, 
+} from "@mui/icons-material";
+
+import {
+	Typography 
+} from "@mui/material";
 
 const alertStyle = {
   backgroundColor: "rgba(0,0,0,0.9)",
