@@ -1,24 +1,33 @@
 import React, { useState, useEffect } from 'react';
 import ReactGA from 'react-ga4';
-import { useTheme } from '@material-ui/core/styles';
+import theme from '../theme.jsx';
 import {Link} from 'react-router-dom';
-import { useAlert } from "react-alert";
-import { Search as SearchIcon, CloudQueue as CloudQueueIcon, Code as CodeIcon } from '@material-ui/icons';
+import { Search as SearchIcon, CloudQueue as CloudQueueIcon, Code as CodeIcon } from '@mui/icons-material';
+import { toast } from 'react-toastify';
 
 //import algoliasearch from 'algoliasearch/lite';
 import algoliasearch from 'algoliasearch';
 import { InstantSearch, connectSearchBox, connectHits } from 'react-instantsearch-dom';
-import { Grid, Paper, TextField, ButtonBase, InputAdornment, Typography, Button, Tooltip} from '@material-ui/core';
+import { 
+	Grid, 
+	Paper, 
+	TextField, 
+	ButtonBase, 
+	InputAdornment, 
+	Typography, 
+	Button, 
+	Tooltip
+} from '@mui/material';
+
 import aa from 'search-insights'
 const searchClient = algoliasearch("JNSS5CFDZZ", "db08e40265e2941b9a7d8f644b6e5240")
 const Appsearch = props => {
-	const { maxRows, showName, showSuggestion, isMobile, globalUrl, parsedXs, newSelectedApp, setNewSelectedApp, defaultSearch, showSearch, ConfiguredHits, userdata, cy, isCreatorPage, actionImageList, setActionImageList}  = props
+	const { maxRows, showName, showSuggestion, isMobile, globalUrl, parsedXs, newSelectedApp, setNewSelectedApp, defaultSearch, showSearch, ConfiguredHits, userdata, cy, isCreatorPage, actionImageList, setActionImageList, setUserSpecialzedApp }  = props
 
-  const isCloud = window.location.host === "localhost:3002" || window.location.host === "soc2.khulnasoft.com";
-  const alert = useAlert();
+    const isCloud = window.location.host === "localhost:3002" || window.location.host === "gsoc2r.io";
 	const rowHandler = maxRows === undefined || maxRows === null ? 50 : maxRows
 	const xs = parsedXs === undefined || parsedXs === null ? 12 : parsedXs
-	const theme = useTheme();
+	//const theme = useTheme();
 	//const [apps, setApps] = React.useState([]);
 	//const [filteredApps, setFilteredApps] = React.useState([]);
 	const [formMail, setFormMail] = React.useState("");
@@ -31,82 +40,6 @@ const Appsearch = props => {
 	const borderRadius = 3
 	window.title = "Gsoc2 | Apps | Find and integration any app"
 
-	const setUserSpecialzedApp = (user, data) => {
-		// var data = newfields]
-		console.log("data value", data)
-		const appData = {"user_id":user,"specialized_apps":[{}]}
-		console.log("User Check for appdata:", user)
-		appData["specialized_apps"][0]["name"] = data["name"]
-		appData["specialized_apps"][0]["image"] = data["image_url"]
-		appData["specialized_apps"][0]["category"] = data["categories"].toString()
-		console.log("AppData:",appData)
-		console.log("setActionImageList",setActionImageList)
-		console.log("actionImageList",actionImageList)
-
-		const finalData = actionImageList.concat(appData["specialized_apps"])
-		appData["specialized_apps"]=finalData
-		fetch(globalUrl + "/api/v1/users/updateuser", {
-		  method: "PUT",
-		  headers: {
-			"Content-Type": "application/json",
-			Accept: "application/json",
-		  },
-		  body: JSON.stringify(appData),
-		  credentials: "include",
-		})
-		  .then((response) => {
-			if (response.status !== 200) {
-			  console.log("Status not 200 for set creator :O!");
-			}
-			alert.success("Sucessfully updated specialzed app.")
-			return response.json();
-		  })
-		  .then((responseJson) => {
-			if (!responseJson.success && responseJson.reason !== undefined) {
-			  alert.error("Failed updating user: " + responseJson.reason);
-			}
-		  })
-		  .catch((error) => {
-			console.log(error);
-		  });
-	  };
-	const submitContact = (email, message) => {
-		const data = {
-			"firstname": "",
-			"lastname": "",
-			"title": "",
-			"companyname": "",
-			"email": email,
-			"phone": "",
-			"message": message,
-		}
-	
-		const errorMessage = "Something went wrong. Please contact gsoc2@soc2.khulnasoft.com directly."
-
-		fetch(globalUrl+"/api/v1/contact", {
-			method: 'POST',
-			headers: {
-				'Content-Type': 'application/json',
-			},
-			body: JSON.stringify(data),
-		})
-		.then(response => response.json())
-		.then(response => {
-			if (response.success === true) {
-				setFormMessage(response.reason)
-				//alert.info("Thanks for submitting!")
-			} else {
-				setFormMessage(errorMessage)
-			}
-
-			setFormMail("")
-			setMessage("")
-    })
-		.catch(error => {
-			setFormMessage(errorMessage)
-    	console.log(error)
-		});
-	}
 
 	// value={currentRefinement}
 	const SearchBox = ({currentRefinement, refine, isSearchStalled} ) => {
@@ -121,14 +54,14 @@ const Appsearch = props => {
 		  <form noValidate action="" role="search">
 				<TextField 
 					fullWidth
-					style={{backgroundColor: theme.palette.inputColor, borderRadius: borderRadius, width: "100%",}} 
+					style={{backgroundColor: "#2F2F2F", borderRadius: borderRadius, width: "100%",}} 
 					InputProps={{
 						style:{
 							color: "white",
 							fontSize: "1em",
 							height: 50,
 						},
-						startAdornment: (
+						endAdornment: (
 							<InputAdornment position="start">
 								<SearchIcon style={{marginLeft: 5}}/>
 							</InputAdornment>
@@ -138,7 +71,8 @@ const Appsearch = props => {
 					type="search"
 					color="primary"
 					defaultValue={defaultSearch}
-					placeholder={`Find ${defaultSearch} Apps...`}
+					// placeholder={`Find ${defaultSearch} Apps...`}
+					placeholder= {defaultSearch ? `${defaultSearch}` : "Search Cases "}
 					id="gsoc2_workflow_search_field"
 					onChange={(event) => {
 						refine(event.currentTarget.value)
@@ -159,9 +93,9 @@ const Appsearch = props => {
 			<Grid container spacing={0} style={{border: "1px solid rgba(255,255,255,0.2)", maxHeight: 250, minHeight: 250, overflowY: "auto", overflowX: "hidden", }}>
 				{hits.map((data, index) => {
 					const paperStyle = {
-						backgroundColor: index === mouseHoverIndex ? "rgba(255,255,255,0.8)" : theme.palette.inputColor,
+						backgroundColor: index === mouseHoverIndex ? "rgba(255,255,255,0.8)" : "#2F2F2F",
 						color: index === mouseHoverIndex ? theme.palette.inputColor : "rgba(255,255,255,0.8)", 
-						border: newSelectedApp.objectID !== data.objectID ? `1px solid rgba(255,255,255,0.2)` : "2px solid #f86a3e", 
+						// border: newSelectedApp.objectID !== data.objectID ? `1px solid rgba(255,255,255,0.2)` : "2px solid #f86a3e", 
 						textAlign: "left",
 						padding: 10,
 						cursor: "pointer",
@@ -178,42 +112,17 @@ const Appsearch = props => {
 
 					counted += 1
 					var parsedname = data.name.valueOf()
-					//for (var key = 0; key < data.name.length; key++) {
-					//	var character = data.name.charAt(key)
-					//	if (character === character.toUpperCase()) {
-					//		//console.log(data.name[key], data.name[key+1])
-					//		if (data.name.charAt(key+1) !== undefined && data.name.charAt(key+1) === data.name.charAt(key+1).toUpperCase()) {
-					//		} else {
-					//			parsedname += " "
-					//		}
-					//	}
-
-					//	parsedname += character
-					//}
-					
 					parsedname = (parsedname.charAt(0).toUpperCase()+parsedname.substring(1)).replaceAll("_", " ")
 
 					return (
 						<Paper key={index} elevation={0} style={paperStyle} onMouseOver={() => {
 							setMouseHoverIndex(index)
-							/*
-							ReactGA.event({
-								category: "app_grid_view",
-								action: `search_bar_click`,
-								label: "",
-							})
-							*/
 						}} onMouseOut={() => {
 							setMouseHoverIndex(-1)
 						}} onClick={() => {
 							if(isCreatorPage === true){
-								console.log("data:",data)
-								console.log("userdata.id",userdata.id)
-								console.log("is creator", isCreatorPage)
-								if (setNewSelectedApp !== undefined) {
-									// setUserSpecialzedApp = data
+								if (setNewSelectedApp !== undefined && setUserSpecialzedApp !== undefined) {
 									setUserSpecialzedApp(userdata.id, data)
-									//setActionImageList(userdata.id, data)
 								}	
 							}
 							if (setNewSelectedApp !== undefined) {
@@ -255,7 +164,7 @@ const Appsearch = props => {
 							}
 						}}>
 							<div style={{display: "flex"}}>
-								<img alt={data.name} src={data.image_url} style={{width: "100%", maxWidth: 30, minWidth: 30, minHeight: 30, maxHeight: 30, display: "block", }} />
+								<img alt={data.name} src={data.image_url} style={{width: "100%", maxWidth: 30, minWidth: 30, minHeight: 30, borderRadius: 40, maxHeight: 30, display: "block", }} />
 								<Typography variant="body1" style={{marginTop: 2, marginLeft: 10, }}>
 									{parsedname}
 								</Typography>
@@ -272,13 +181,8 @@ const Appsearch = props => {
 	const CustomHits = connectHits(InputHits)
 
 	return (
-		<div style={{width: "100%", textAlign: "center", position: "relative", height: "100%",}}>
+		<div style={{width: isMobile ? null : 287, height: 295, padding: "16px 16px 267px 16px", alignItems: "center", gap: 138,}}>
 			<InstantSearch searchClient={searchClient} indexName="appsearch">
-				{/* showSearch === false ? null : 
-					<div style={{maxWidth: 450, margin: "auto", }}>
-						<CustomSearchBox />
-					</div>
-				*/}
 				<div style={{maxWidth: 450, margin: "auto", }}>
 					<CustomSearchBox />
 				</div>
