@@ -12,12 +12,12 @@ import {
   CircularProgress,
 	Tooltip,
 	Button,
-} from "@material-ui/core";
+} from "@mui/material";
 
 import {
 	Close as CloseIcon,
   Delete as DeleteIcon,
-} from "@material-ui/icons";
+} from "@mui/icons-material";
 
 const AppSearchPopout = (props) => {
   const { 
