@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import theme from '../theme.jsx';
 import { v4 as uuidv4 } from "uuid";
-
+import { toast } from 'react-toastify';
 
 import {
   Button,
@@ -13,11 +13,11 @@ import {
 	DialogTitle,
 	DialogContent,
 	Typography,
-} from "@material-ui/core";
+} from "@mui/material";
 
 import {
   LockOpen as LockOpenIcon,
-} from "@material-ui/icons";
+} from "@mui/icons-material";
 
 const AuthenticationData = (props) => {
   const {
@@ -54,22 +54,22 @@ const AuthenticationData = (props) => {
       })
       .then((responseJson) => {
         if (!responseJson.success) {
-          alert.error("Failed to set app auth: " + responseJson.reason);
+          toast("Failed to set app auth: " + responseJson.reason);
         } else {
-					if (getAppAuthentication !== undefined) {
-          	getAppAuthentication()
-					}
+			if (getAppAuthentication !== undefined) {
+          		getAppAuthentication()
+			}
 
-					if (setAuthenticationModalOpen !== undefined) {
-						setAuthenticationModalOpen(false)
-					}
+			if (setAuthenticationModalOpen !== undefined) {
+				setAuthenticationModalOpen(false)
+			}
 
           // Needs a refresh with the new authentication..
-          //alert.success("Successfully saved new app auth")
+          //toast("Successfully saved new app auth")
         }
       })
       .catch((error) => {
-        //alert.error(error.toString());
+        //toast(error.toString());
         console.log("New auth error: ", error.toString());
       });
   }
@@ -146,7 +146,7 @@ const AuthenticationData = (props) => {
 							selectedApp.authentication.parameters[key].name
 						] = "false";
 					} else {
-						alert.info(
+						toast(
 							"Field " +
 								selectedApp.authentication.parameters[key].name +
 								" can't be empty"
@@ -210,7 +210,7 @@ const AuthenticationData = (props) => {
 				<a
 					target="_blank"
 					rel="noopener noreferrer"
-					href="https://soc2.khulnasoft.com/docs/apps#authentication"
+					href="https://gsoc2r.io/docs/apps#authentication"
 					style={{ textDecoration: "none", color: "#f85a3e" }}
 				>
 					What is app authentication?
