@@ -1,5 +1,7 @@
 import React, { useState, useEffect } from "react";
 import theme from "../theme.jsx";
+import { toast } from 'react-toastify';
+
 import {
     Tooltip,
     Divider,
@@ -15,8 +17,7 @@ import {
     Dialog,
     DialogTitle,
     DialogActions,
-} from "@material-ui/core";
-import { useAlert } from "react-alert";
+} from "@mui/material";
 
 import {
     Edit as EditIcon,
@@ -40,8 +41,7 @@ import {
     Business as BusinessIcon,
     Visibility as VisibilityIcon,
     VisibilityOff as VisibilityOffIcon,
-} from "@material-ui/icons";
-import data from "./frameworkStyle.jsx";
+} from "@mui/icons-material";
 
 const scrollStyle1 = {
     height: 100,
@@ -73,7 +73,7 @@ const CacheView = (props) => {
     const [dataValue, setDataValue] = React.useState({});
     const [editCache, setEditCache] = React.useState(false);
     const [show, setShow] = useState({});
-    const alert = useAlert();
+
     useEffect(() => {
         listOrgCache(orgId);
         console.log("orgid", orgId);
@@ -106,7 +106,7 @@ const CacheView = (props) => {
                 }
             })
             .catch((error) => {
-                alert.error(error.toString());
+                toast(error.toString());
             });
     };
 
@@ -149,7 +149,7 @@ const CacheView = (props) => {
 
 
     const deleteCache = (orgId, key) => {
-        alert.info("Attempting to delete Cache");
+        toast("Attempting to delete Cache");
         fetch(globalUrl + `/api/v1/orgs/${orgId}/cache/${key}`, {
             method: "DELETE",
             headers: {
@@ -159,16 +159,16 @@ const CacheView = (props) => {
         })
             .then((response) => {
                 if (response.status === 200) {
-                    alert.success("Successfully deleted Cache");
+                    toast("Successfully deleted Cache");
                     setTimeout(() => {
                         listOrgCache(orgId);
                     }, 1000);
                 } else {
-                    alert.error("Failed deleting Cache. Does it still exist?");
+                    toast("Failed deleting Cache. Does it still exist?");
                 }
             })
             .catch((error) => {
-                alert.error(error.toString());
+                toast(error.toString());
             });
     };
 
@@ -198,12 +198,12 @@ const CacheView = (props) => {
             })
             .then((responseJson) => {
                 setAddCache(responseJson);
-                alert.success("Cache Edited Successfully!");
+                toast("Cache Edited Successfully!");
                 listOrgCache(orgId);
                 setModalOpen(false);
             })
             .catch((error) => {
-                alert.error(error.toString());
+                toast(error.toString());
             });
     };
 
@@ -232,12 +232,12 @@ const CacheView = (props) => {
             })
             .then((responseJson) => {
                 setAddCache(responseJson);
-                alert.success("New Cache Added Successfully!");
+                toast("New Cache Added Successfully!");
                 listOrgCache(orgId);
                 setModalOpen(false);
             })
             .catch((error) => {
-                alert.error(error.toString());
+                toast(error.toString());
             });
     };
 
@@ -339,7 +339,7 @@ const CacheView = (props) => {
             <div style={{ marginTop: 20, marginBottom: 20 }}>
                 <h2 style={{ display: "inline" }}>Gsoc2 Datastore</h2>
                 <span style={{ marginLeft: 25 }}>
-                    Datastore is a key-value store for storing data that can be used cross-workflow.&nbsp;
+                    Datastore is a permanent key-value database for storing data that can be used cross-workflow. You can store anything from lists of IPs to complex configurations.&nbsp;
                     <a
                         target="_blank"
                         rel="noopener noreferrer"
