@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from "react";
+import { toast } from 'react-toastify';
 
 import {
   IconButton,
@@ -15,7 +16,7 @@ import {
 	Divider,
 	Select,
 	MenuItem,
-} from "@material-ui/core";
+} from "@mui/material";
 
 import {
   OpenInNew as OpenInNewIcon,
@@ -27,9 +28,9 @@ import {
   Publish as PublishIcon,
 	Clear as ClearIcon,
 	Add as AddIcon,
-} from "@material-ui/icons";
+} from "@mui/icons-material";
 
-import { useAlert } from "react-alert";
+//import { useAlert 
 import Dropzone from "../components/Dropzone.jsx";
 import CodeEditor from "../components/Gsoc2CodeEditor.jsx";
 import theme from "../theme.jsx";
@@ -45,7 +46,7 @@ const Files = (props) => {
   const [openEditor, setOpenEditor] = React.useState(false);
   const [renderTextBox, setRenderTextBox] = React.useState(false);
 
-  const alert = useAlert();
+  //const alert = useAlert();
   const allowedFileTypes = ["txt", "py", "yaml", "yml","json", "html", "js", "csv", "log"]
   var upload = "";
 
@@ -113,7 +114,7 @@ const Files = (props) => {
         }
       })
       .catch((error) => {
-        alert.error(error.toString());
+        toast(error.toString());
       });
   };
 
@@ -139,12 +140,12 @@ const Files = (props) => {
       })
       .then((responseJson) => {
         if (responseJson.success) {
-          alert.info("Successfully deleted file " + file.name);
+          toast("Successfully deleted file " + file.name);
         } else if (
           responseJson.reason !== undefined &&
           responseJson.reason !== null
         ) {
-          alert.error("Failed to delete file: " + responseJson.reason);
+          toast("Failed to delete file: " + responseJson.reason);
         }
         setTimeout(() => {
           getFiles();
@@ -153,7 +154,7 @@ const Files = (props) => {
         console.log(responseJson);
       })
       .catch((error) => {
-        alert.error(error.toString());
+        toast(error.toString());
       });
   };
 
@@ -178,7 +179,7 @@ const Files = (props) => {
           // console.log("respdata type ->", typeof(respdata));
         
         if (respdata.length === 0) {
-          alert.error("Failed getting file. Is it deleted?");
+          toast("Failed getting file. Is it deleted?");
           return;
         }
         return respdata
@@ -189,7 +190,7 @@ const Files = (props) => {
       //console.log("filecontent state ",fileContent);
       })
       .catch((error) => {
-        alert.error(error.toString());
+        toast(error.toString());
       });
   };
 
@@ -210,7 +211,7 @@ const Files = (props) => {
 		})
       .then((respdata) => {
         if (respdata.length === 0) {
-          alert.error("Failed getting file. Is it deleted?");
+          toast("Failed getting file. Is it deleted?");
           return;
         }
 
@@ -249,7 +250,7 @@ const Files = (props) => {
         //setSchedules(responseJson)
       })
       .catch((error) => {
-        alert.error(error.toString());
+        toast(error.toString());
       });
   };
 
@@ -291,11 +292,11 @@ const Files = (props) => {
         if (responseJson.success === true) {
           handleFileUpload(responseJson.id, file);
         } else {
-          alert.error("Failed to upload file ", filename);
+          toast("Failed to upload file ", filename);
         }
       })
       .catch((error) => {
-        alert.error("Failed to upload file ", filename);
+        toast("Failed to upload file ", filename);
         console.log(error.toString());
       });
   };
@@ -312,7 +313,7 @@ const Files = (props) => {
       .then((response) => {
         if (response.status !== 200 && response.status !== 201) {
           console.log("Status not 200 for apps :O!");
-          alert.error("File was created, but failed to upload.");
+          toast("File was created, but failed to upload.");
           return;
         }
 
@@ -323,7 +324,7 @@ const Files = (props) => {
         //setFiles(responseJson)
       })
       .catch((error) => {
-        alert.error(error.toString());
+        toast(error.toString());
       });
   };
 
@@ -364,7 +365,7 @@ const Files = (props) => {
     const files = isDropzone ? e.dataTransfer.files : e.target.files;
 
     //const reader = new FileReader();
-    //alert.info("Starting fileupload")
+    //toast("Starting fileupload")
     uploadFiles(files);
   };
 
@@ -381,11 +382,11 @@ const Files = (props) => {
 				<div style={{ marginTop: 20, marginBottom: 20 }}>
 					<h2 style={{ display: "inline" }}>Files</h2>
 					<span style={{ marginLeft: 25 }}>
-						Files from Workflows.{" "}
+						Files from Workflows are a way to store as well as edit files.{" "}
 						<a
 							target="_blank"
 							rel="noopener noreferrer"
-							href="https://soc2.khulnasoft.com/docs/organizations#files"
+							href="https://gsoc2r.io/docs/organizations#files"
 							style={{ textDecoration: "none", color: "#f85a3e" }}
 						>
 							Learn more
@@ -681,7 +682,7 @@ const Files = (props) => {
 									<ListItemText
 										primary=<span style={{ display:"inline"}}>
 											<Tooltip
-												title={`Edit File (${allowedFileTypes.join(", ")})`}
+												title={`Edit File (${allowedFileTypes.join(", ")}). Max size 2MB`}
 												style={{}}
 												aria-label={"Edit"}
 											>
@@ -742,7 +743,7 @@ const Files = (props) => {
 														) {
 															const clipboard = navigator.clipboard;
 															if (clipboard === undefined) {
-																alert.error(
+																toast(
 																	"Can only copy over HTTPS (port 3443)"
 																);
 																return;
@@ -758,7 +759,7 @@ const Files = (props) => {
 															/* Copy the text inside the text field */
 															document.execCommand("copy");
 
-															alert.info(file.id + " copied to clipboard");
+															toast(file.id + " copied to clipboard");
 														}
 													}}
 												>
