@@ -1,6 +1,6 @@
 /* eslint-disable react/no-multi-comp */
 import React, { useState } from "react";
-import { makeStyles } from "@material-ui/styles";
+import { makeStyles } from "@mui/styles";
 
 import {
   CircularProgress,
@@ -8,7 +8,7 @@ import {
   Button,
   Paper,
   Typography,
-} from "@material-ui/core";
+} from "@mui/material";
 
 const bodyDivStyle = {
   margin: "auto",
