@@ -1,7 +1,8 @@
 /* eslint-disable react/no-multi-comp */
 import React, { useState, useEffect } from "react";
-import { makeStyles } from "@material-ui/styles";
+import { makeStyles } from "@mui/styles";
 import { useInterval } from "react-powerhooks";
+import theme from '../theme.jsx';
 
 import {
   CircularProgress,
@@ -9,9 +10,8 @@ import {
   Button,
   Paper,
   Typography,
-} from "@material-ui/core";
+} from "@mui/material";
 
-import { useTheme } from "@material-ui/core/styles";
 import { useNavigate } from "react-router-dom";
 
 const hrefStyle = {
@@ -42,7 +42,6 @@ const LoginDialog = (props) => {
     checkLogin,
   } = props;
 
-  const theme = useTheme();
 	let navigate = useNavigate();
   const classes = useStyles();
 
@@ -306,7 +305,6 @@ const LoginDialog = (props) => {
                 paddingBottom: "30px",
                 paddingTop: "30px",
                 position: "relative",
-                backgroundColor: theme.palette.inputColor,
                 textAlign: "left",
                 marginTop: 15,
               }}
@@ -332,20 +330,28 @@ const LoginDialog = (props) => {
                 variant="body2"
                 style={{ marginBottom: 20, color: "white" }}
               >
-                <b>1.</b> Make sure gsoc2-database folder has correct access:{" "}
+                <b>1.</b> Make sure gsoc2-database folder has correct access, and that you have a minimum of <b>2Gb of RAM available</b>:{" "}
                 <br />
                 <br />
-                sudo chown 1000:1000 -R gsoc2-database
+                sudo chown -R 1000:1000 gsoc2-database
               </Typography>
-
               <Typography
                 variant="body2"
                 style={{ marginBottom: 20, color: "white" }}
               >
-                <b>2</b>. Restart docker-compose:
+                <b>2.</b> Disable memory swap on the host:
                 <br />
                 <br />
-                sudo docker-compose restart
+				sudo swapoff -a
+              </Typography>
+              <Typography
+                variant="body2"
+                style={{ marginBottom: 20, color: "white" }}
+              >
+                <b>3</b>. Restart the database:
+                <br />
+                <br />
+                sudo docker restart gsoc2-opensearch
               </Typography>
             </Paper>
             <Typography
@@ -514,7 +520,7 @@ const LoginDialog = (props) => {
 				//id="sso_button"
     		const ssoBtn = document.getElementById("sso_button");
 				if (ssoBtn !== undefined && ssoBtn !== null) {
-					console.log("SSO BTN: ", ssoBtn)
+					//console.log("SSO BTN: ", ssoBtn)
 					const cursearch = typeof window === "undefined" || window.location === undefined ? "" : window.location.search;
 					var tmpView = new URLSearchParams(cursearch).get("autologin");
 					if (tmpView !== undefined && tmpView !== null) {
@@ -528,7 +534,7 @@ const LoginDialog = (props) => {
 		}, 200);
 	}, [ssoUrl])
 
-  return <div>{loadedCheck}</div>;
+  return <div style={{paddingBottom: 150, }}>{loadedCheck}</div>;
 };
 
 export default LoginDialog;
