@@ -31,7 +31,7 @@ def parse_data(data):
             "contact": {
                 "name": "@gsoc2likeme",
                 "url": "https://twitter.com/gsoc2likeme",
-                "email": "gsoc2@soc2.khulnasoft.com"
+                "email": "gsoc2@gsoc2r.io"
             }
         },
         "paths": {},
