@@ -300,7 +300,7 @@ from walkoff_app_sdk.app_base import AppBase
 
 class %s(AppBase):
     """
-    Autogenerated class by Gsoc2
+    Autogenerated class by Gsoc2r
     """
     
     __version__ = "%s"
@@ -345,7 +345,7 @@ func dumpApi(basePath string, api WorkflowApp) error {
 }
 
 func main() {
-	data := []byte(`{"swagger":"3.0","info":{"title":"hi","description":"you","version":"1.0"},"servers":[{"url":"https://soc2.khulnasoft.com/api/v1"}],"host":"soc2.khulnasoft.com","basePath":"/api/v1","schemes":["https:"],"paths":{"/workflows":{"get":{"responses":{"default":{"description":"default","schema":{}}},"summary":"Get workflows","description":"Get workflows","parameters":[]}},"/workflows/{id}":{"get":{"responses":{"default":{"description":"default","schema":{}}},"summary":"Get workflow","description":"Get workflow","parameters":[{"in":"query","name":"forgetme","description":"Generated by soc2.khulnasoft.com OpenAPI","required":true,"schema":{"type":"string"}},{"in":"query","name":"anotherone","description":"Generated by soc2.khulnasoft.com OpenAPI","required":false,"schema":{"type":"string"}},{"in":"query","name":"hi","description":"Generated by soc2.khulnasoft.com OpenAPI","required":true,"schema":{"type":"string"}},{"in":"path","name":"id","description":"Generated by soc2.khulnasoft.com OpenAPI","required":true,"schema":{"type":"string"}}]}}},"securityDefinitions":{}}`)
+	data := []byte(`{"swagger":"3.0","info":{"title":"hi","description":"you","version":"1.0"},"servers":[{"url":"https://gsoc2r.io/api/v1"}],"host":"gsoc2r.io","basePath":"/api/v1","schemes":["https:"],"paths":{"/workflows":{"get":{"responses":{"default":{"description":"default","schema":{}}},"summary":"Get workflows","description":"Get workflows","parameters":[]}},"/workflows/{id}":{"get":{"responses":{"default":{"description":"default","schema":{}}},"summary":"Get workflow","description":"Get workflow","parameters":[{"in":"query","name":"forgetme","description":"Generated by gsoc2r.io OpenAPI","required":true,"schema":{"type":"string"}},{"in":"query","name":"anotherone","description":"Generated by gsoc2r.io OpenAPI","required":false,"schema":{"type":"string"}},{"in":"query","name":"hi","description":"Generated by gsoc2r.io OpenAPI","required":true,"schema":{"type":"string"}},{"in":"path","name":"id","description":"Generated by gsoc2r.io OpenAPI","required":true,"schema":{"type":"string"}}]}}},"securityDefinitions":{}}`)
 
 	hasher := md5.New()
 	hasher.Write(data)
