@@ -27,7 +27,7 @@ import (
 	"gopkg.in/yaml.v2"
 )
 
-var bucketName = "gsoc2.appspot.com"
+var bucketName = "gsoc2r.appspot.com"
 
 type WorkflowApp struct {
 	Name        string `json:"name" yaml:"name" required:true datastore:"name"`
@@ -336,7 +336,7 @@ from walkoff_app_sdk.app_base import AppBase
 
 class %s(AppBase):
     """
-    Autogenerated class by Gsoc2
+    Autogenerated class by Gsoc2r
     """
     
     __version__ = "%s"
@@ -381,7 +381,7 @@ from walkoff_app_sdk.app_base import AppBase
 
 class %s(AppBase):
     """
-    Autogenerated class by Gsoc2
+    Autogenerated class by Gsoc2r
     """
     
     __version__ = "%s"
@@ -451,7 +451,7 @@ func dumpApi(basePath string, api WorkflowApp) error {
 }
 
 func main() {
-	data := []byte(`{"swagger":"3.0","info":{"title":"hi","description":"you","version":"1.0"},"servers":[{"url":"https://soc2.khulnasoft.com/api/v1"}],"host":"soc2.khulnasoft.com","basePath":"/api/v1","schemes":["https:"],"paths":{"/workflows":{"get":{"responses":{"default":{"description":"default","schema":{}}},"summary":"Get workflows","description":"Get workflows","parameters":[]}},"/workflows/{id}":{"get":{"responses":{"default":{"description":"default","schema":{}}},"summary":"Get workflow","description":"Get workflow","parameters":[{"in":"query","name":"forgetme","description":"Generated by soc2.khulnasoft.com OpenAPI","required":true,"schema":{"type":"string"}},{"in":"query","name":"anotherone","description":"Generated by soc2.khulnasoft.com OpenAPI","required":false,"schema":{"type":"string"}},{"in":"query","name":"hi","description":"Generated by soc2.khulnasoft.com OpenAPI","required":true,"schema":{"type":"string"}},{"in":"path","name":"id","description":"Generated by soc2.khulnasoft.com OpenAPI","required":true,"schema":{"type":"string"}}]}}},"securityDefinitions":{}}`)
+	data := []byte(`{"swagger":"3.0","info":{"title":"hi","description":"you","version":"1.0"},"servers":[{"url":"https://gsoc2r.io/api/v1"}],"host":"gsoc2r.io","basePath":"/api/v1","schemes":["https:"],"paths":{"/workflows":{"get":{"responses":{"default":{"description":"default","schema":{}}},"summary":"Get workflows","description":"Get workflows","parameters":[]}},"/workflows/{id}":{"get":{"responses":{"default":{"description":"default","schema":{}}},"summary":"Get workflow","description":"Get workflow","parameters":[{"in":"query","name":"forgetme","description":"Generated by gsoc2r.io OpenAPI","required":true,"schema":{"type":"string"}},{"in":"query","name":"anotherone","description":"Generated by gsoc2r.io OpenAPI","required":false,"schema":{"type":"string"}},{"in":"query","name":"hi","description":"Generated by gsoc2r.io OpenAPI","required":true,"schema":{"type":"string"}},{"in":"path","name":"id","description":"Generated by gsoc2r.io OpenAPI","required":true,"schema":{"type":"string"}}]}}},"securityDefinitions":{}}`)
 
 	ctx := context.Background()
 	client, err := storage.NewClient(ctx)
