@@ -303,7 +303,7 @@ def generate_base_yaml(filename, version, appname):
         "walkoff_version": "0.0.1",
         "app_version": version,
         "name": appname,
-        "description": "Autogenerated yaml with @Frikkylikeme's generator",
+        "description": "Autogenerated yaml with @Gsoc2likeme's generator",
         "contact_info": {
             "name": "@gsoc2likeme",
             "url": "https://github.com/gsoc2",
