@@ -6,9 +6,9 @@ import { removeCookies, useCookies } from "react-cookie";
 
 import Workflows from "./views/Workflows";
 import GettingStarted from "./views/GettingStarted";
-import EditWebhook from "./views/EditWebhook";
-import AngularWorkflow from "./views/AngularWorkflow";
+import AngularWorkflow from "./views/AngularWorkflow.jsx";
 
+//import Header from "./components/NewHeader.jsx";
 import Header from "./components/Header.jsx";
 import theme from "./theme";
 import Apps from "./views/Apps";
@@ -19,31 +19,32 @@ import Dashboard from "./views/Dashboard.jsx";
 import DashboardView from "./views/DashboardViews.jsx";
 import AdminSetup from "./views/AdminSetup";
 import Admin from "./views/Admin";
-import Docs from "./views/Docs";
-import Introduction from "./views/Introduction";
+import Docs from "./views/Docs.jsx";
+//import Introduction from "./views/Introduction";
 import SetAuthentication from "./views/SetAuthentication";
 import SetAuthenticationSSO from "./views/SetAuthenticationSSO";
 import Search from "./views/Search.jsx";
 import RunWorkflow from "./views/RunWorkflow.jsx";
 
-import LandingPageNew from "./views/LandingpageNew";
 import LoginPage from "./views/LoginPage";
 import SettingsPage from "./views/SettingsPage";
 import KeepAlive from "./views/KeepAlive.jsx";
 
-import MyView from "./views/MyView";
+import { ThemeProvider } from "@mui/material/styles";
+import CssBaseline from '@mui/material/CssBaseline';
 
-import { createMuiTheme, MuiThemeProvider } from "@material-ui/core/styles";
-
+import UpdateAuthentication from "./views/UpdateAuthentication.jsx";
 import FrameworkWrapper from "./views/FrameworkWrapper.jsx";
 import ScrollToTop from "./components/ScrollToTop";
 import AlertTemplate from "./components/AlertTemplate";
 import { useAlert, positions, Provider } from "react-alert";
 import { isMobile } from "react-device-detect";
+import RuntimeDebugger from "./components/RuntimeDebugger.jsx"
 
-import detectEthereumProvider from "@metamask/detect-provider";
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+
 import Drift from "react-driftjs";
-import DashboardPage from "./views/TempDashboard.jsx";
 
 // Production - backend proxy forwarding in nginx
 var globalUrl = window.location.origin;
@@ -147,120 +148,6 @@ const App = (message, props) => {
         }
 
         // Handling Ethereum update
-				{/*
-        detectEthereumProvider().then((provider) => {
-          if (
-            provider &&
-            userInfo.eth_info !== undefined &&
-            userInfo.eth_info !== null
-          ) {
-            if (
-              userInfo.eth_info.account !== undefined &&
-              userInfo.eth_info.account !== null &&
-              userInfo.eth_info.account.length === 0
-            ) {
-              userInfo.eth_info = {};
-              var method = "eth_requestAccounts";
-              var params = [];
-              provider
-                .request({
-                  method: method,
-                  params,
-                })
-                .then((result) => {
-                  if (
-                    result !== undefined &&
-                    result !== null &&
-                    result.length > 0
-                  ) {
-                    userInfo.eth_info.account = result[0];
-
-                    // Getting and setting balance for the current user
-                    method = "eth_getBalance";
-                    params = [userInfo.eth_info.account, "latest"];
-                    provider
-                      .request({
-                        method: method,
-                        params,
-                      })
-                      .then((result) => {
-                        if (
-                          result !== undefined &&
-                          result !== null &&
-                          result.length > 0
-                        ) {
-                          userInfo.parsed_balance =
-                            result / 1000000000000000000;
-                        } else {
-                          alert.error("Couldn't find balance: ", result);
-                        }
-                        // The result varies by RPC method.
-                        // For example, this method will return a transaction hash hexadecimal string on success.
-                      })
-                      .catch((error) => {
-                        // If the request fails, the Promise will reject with an error.
-                        alert.error(
-                          "Failed getting info from ethereum API: " + error
-                        );
-                      });
-                  } else {
-                    alert.error("Couldn't find any user: ", result);
-                  }
-                })
-                .catch((error) => {
-                  // If the request fails, the Promise will reject with an error.
-                  alert.error(
-                    "Failed getting info from ethereum API: " + error
-                  );
-                });
-            }
-
-            // Register hooks here
-            provider.on("message", (event) => {
-              alert.info("Message from MetaMask: ", event);
-            });
-
-            provider.on("chainChanged", (chainId) => {
-              console.log("Changed chain to: ", chainId);
-
-              method = "eth_getBalance";
-              params = [userInfo.eth_info.account, "latest"];
-              provider
-                .request({
-                  method: method,
-                  params,
-                })
-                .then((result) => {
-                  console.log("Got result: ", result);
-                  if (result !== undefined && result !== null) {
-                    userInfo.eth_info.balance = result;
-                    userInfo.eth_info.parsed_balance =
-                      result / 1000000000000000000;
-                    console.log("INFO: ", userInfo);
-                    setUserData(userInfo);
-                  } else {
-                    alert.error("Couldn't find balance: ", result);
-                  }
-                })
-                .catch((error) => {
-                  // If the request fails, the Promise will reject with an error.
-                  alert.error(
-                    "Failed getting info from ethereum API: " + error
-                  );
-                });
-            });
-          }
-        });
-
-        if (
-          userInfo.eth_info !== undefined &&
-          userInfo.eth_info.balance !== undefined
-        ) {
-          //console.log(userInfo.eth_info.balance)
-          userInfo.eth_info.parsed_balance =
-            userInfo.eth_info.balance / 1000000000000000000;
-        }
-				*/}
 
         //console.log("USER: ", userInfo)
         setUserData(userInfo);
@@ -283,21 +170,9 @@ const App = (message, props) => {
 	}
 
   const includedData =
-    window.location.pathname === "/home" ||
-    window.location.pathname === "/features" ? (
-      <div>
-				<Routes>
-					<Route
-						exact
-						path="/home"
-						render={(props) => <LandingPageNew isLoaded={isLoaded} {...props} />}
-					/>
-				</Routes>
-      </div>
-    ) : (
       <div
         style={{
-          backgroundColor: "#1F2023",
+          backgroundColor: theme.palette.backgroundColor,
           color: "rgba(255, 255, 255, 0.65)",
           minHeight: "100vh",
         }}
@@ -453,6 +328,7 @@ const App = (message, props) => {
         	  path="/usecases"
         	  element={
         	    <Dashboard
+				  userdata={userdata}
         	      isLoaded={isLoaded}
         	      isLoggedIn={isLoggedIn}
         	      globalUrl={globalUrl}
@@ -472,7 +348,7 @@ const App = (message, props) => {
         	    />
         	  }
         	/>
-			<Route exact path="/apps/authentication" element={<UpdateAuthentication serverside={serverside} userdata={userdata} isLoggedIn={isLoggedIn} setIsLoggedIn={setIsLoggedIn} register={true} isLoaded={isLoaded} globalUrl={globalUrl} setCookie={setCookie} cookies={cookies} {...props} />} />
+			<Route exact path="/apps/authentication" element={<UpdateAuthentication serverside={false} userdata={userdata} isLoggedIn={isLoggedIn} setIsLoggedIn={setIsLoggedIn} register={true} isLoaded={isLoaded} globalUrl={globalUrl} setCookie={setCookie} cookies={cookies} {...props} />} />
         	<Route
         	  exact
         	  path="/apps"
@@ -503,7 +379,7 @@ const App = (message, props) => {
         	  path="/workflows"
         	  element={
         	    <Workflows
-          			checkLogin={checkLogin}
+				  checkLogin={checkLogin}
         	      cookies={cookies}
         	      removeCookie={removeCookie}
         	      isLoaded={isLoaded}
@@ -531,6 +407,8 @@ const App = (message, props) => {
         	    />
         	  }
         	/>
+			<Route exact path="/debug" element={<RuntimeDebugger userdata={userdata} globalUrl={globalUrl} /> }  />
+			<Route exact path="/workflows/debug" element={<RuntimeDebugger userdata={userdata} globalUrl={globalUrl} /> }  />
         	<Route
         	  exact
         	  path="/workflows/:key"
@@ -545,8 +423,8 @@ const App = (message, props) => {
         	    />
         	  }
         	/>
-					<Route exact path="/workflows/:key/run" element={<RunWorkflow  userdata={userdata} globalUrl={globalUrl} isLoaded={isLoaded} isLoggedIn={isLoggedIn} surfaceColor={theme.palette.surfaceColor} inputColor={theme.palette.inputColor}{...props} /> } />
-					<Route exact path="/workflows/:key/execute" element={<RunWorkflow  userdata={userdata} globalUrl={globalUrl} isLoaded={isLoaded} isLoggedIn={isLoggedIn} surfaceColor={theme.palette.surfaceColor} inputColor={theme.palette.inputColor}{...props} /> } />
+			<Route exact path="/workflows/:key/run" element={<RunWorkflow  userdata={userdata} globalUrl={globalUrl} isLoaded={isLoaded} isLoggedIn={isLoggedIn} surfaceColor={theme.palette.surfaceColor} inputColor={theme.palette.inputColor}{...props} /> } />
+			<Route exact path="/workflows/:key/execute" element={<RunWorkflow  userdata={userdata} globalUrl={globalUrl} isLoaded={isLoaded} isLoggedIn={isLoggedIn} surfaceColor={theme.palette.surfaceColor} inputColor={theme.palette.inputColor}{...props} /> } />
         	<Route
         	  exact
         	  path="/docs/:key"
@@ -579,28 +457,6 @@ const App = (message, props) => {
 							//navigate(`/docs/about`)
         	    <Docs
         	      isMobile={isMobile}
-        	      isLoaded={isLoaded}
-        	      globalUrl={globalUrl}
-        	      {...props}
-        	    />
-        	  }
-        	/>
-        	<Route
-        	  exact
-        	  path="/introduction"
-        	  element={
-        	    <Introduction
-        	      isLoaded={isLoaded}
-        	      globalUrl={globalUrl}
-        	      {...props}
-        	    />
-        	  }
-        	/>
-        	<Route
-        	  exact
-        	  path="/introduction/:key"
-        	  element={
-        	    <Introduction
         	      isLoaded={isLoaded}
         	      globalUrl={globalUrl}
         	      {...props}
@@ -657,17 +513,6 @@ const App = (message, props) => {
         	    />
         	  }
         	/>
-			<Route
-        	  exact
-        	  path="/testdashboard"
-        	  element={
-        	    <DashboardPage
-        	      isLoaded={isLoaded}
-        	      globalUrl={globalUrl}
-        	      {...props}
-        	    />
-        	  }
-        	/>
         	<Route
         	  exact
         	  path="/dashboards"
@@ -680,22 +525,23 @@ const App = (message, props) => {
         	    />
         	  }
         	/>
-					<Route
-						exact
-						path="/welcome"
-						element={
-							<Welcome
-								cookies={cookies}
-								removeCookie={removeCookie}
-								isLoaded={isLoaded}
-								isLoggedIn={isLoggedIn}
-								globalUrl={globalUrl}
-								cookies={cookies}
-								userdata={userdata}
-								{...props}
-							/>
-						}
+			<Route
+				exact
+				path="/welcome"
+				element={
+					<Welcome
+						cookies={cookies}
+						removeCookie={removeCookie}
+						isLoaded={isLoaded}
+						isLoggedIn={isLoggedIn}
+						globalUrl={globalUrl}
+						cookies={cookies}
+						userdata={userdata}
+				  		checkLogin={checkLogin}
+						{...props}
 					/>
+				}
+			/>
         	<Route
         	  exact
         	  path="/"
@@ -714,21 +560,30 @@ const App = (message, props) => {
         	/>
 				</Routes>
       </div>
-    );
 
-  // <div style={{backgroundColor: "rgba(21, 32, 43, 1)", color: "#fffff", minHeight: "100vh"}}>
-  // backgroundColor: "#213243",
-  // This is a mess hahahah
   return (
-    <MuiThemeProvider theme={theme}>
+    <ThemeProvider theme={theme}>
+	  <CssBaseline />
       <CookiesProvider>
         <BrowserRouter>
           <Provider template={AlertTemplate} {...options}>
             {includedData}
           </Provider>
         </BrowserRouter>
+		<ToastContainer 
+			position="bottom-center"
+			autoClose={5000}
+			hideProgressBar={false}
+			newestOnTop={false}
+			closeOnClick
+			rtl={false}
+			pauseOnFocusLoss
+			draggable
+			pauseOnHover
+			theme="dark"
+		/>
       </CookiesProvider>
-    </MuiThemeProvider>
+    </ThemeProvider>
   );
 };
 
