@@ -1,12 +1,16 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
+import { createRoot } from "react-dom/client";
 import App from "./App";
-import * as serviceWorker from "./serviceWorker";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+//import "./index.css";
+//import reportWebVitals from "./reportWebVitals";
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.unregister();
+const rootElement = document.getElementById("root");
+const root = createRoot(rootElement);
+root.render(
+  <React.Fragment>
+    <App />
+  </React.Fragment>
+);
+
+//reportWebVitals();
